drivers: regulator: Update API to match other classes
The regulator drivers were the last to take the device and ID as
separate parameters. Change them to use a single handle parameter, and
remove the extra code for iterating over an array of regulator IDs.

The goal is to support boards where the CPU and DRAM regulators are
supplied by different devices.

Signed-off-by: Samuel Holland <samuel@sholland.org>