common: Disable the DRAM power supply during shutdown
DRAM contents are not meant to be preserved once the system is shut
down. After suspending the DRAM controller, the firmware is free to
remove power from the DRAM chips.

Power must be restored before resetting the SoC, because (at least
on 64-bit platforms), there is no driver that can re-enable the DRAM
supply before DRAM init is run.

Signed-off-by: Samuel Holland <samuel@sholland.org>