drivers: ccu: Split updates to PLL_CTRL_REG1
Depending on which pads the 24 MHz crystal is connected to, the
PLL_CTRL_REG1_CRYSTAL_EN bit may not have any effect. In that case,
there is no need for the extra delays or register writes.

Currently, the helper for writing PLL_CTRL_REG1 does everything at
once in an attempt to reduce code size. However, we need to split the
first and second writes so one can be compiled out. In fact, writing
the code this way actually produces a smaller binary.

Signed-off-by: Samuel Holland <samuel@sholland.org>